% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_main.R
\name{el_main_sec}
\alias{el_main_sec}
\title{Title}
\usage{
el_main_sec(..., .el_bg_color = NULL, .el_is_grid = TRUE)
}
\arguments{
\item{...}{NULL}

\item{.el_bg_color}{NULL}

\item{.el_is_grid}{NULL}
}
\description{
Title
}
\examples{
NULL
}
