% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_formatters.R
\name{format_en_toc}
\alias{format_en_toc}
\title{format_en_toc}
\usage{
format_en_toc(.x)
}
\arguments{
\item{.x}{a named list}
}
\value{
a HTML string
}
\description{
format_en_toc
}
\examples{
FALSE
}
