% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_formatters.R
\name{format_fa_list}
\alias{format_fa_list}
\title{format_fa_list}
\usage{
format_fa_list(.x)
}
\arguments{
\item{.x}{a named list}
}
\value{
a HTML string
}
\description{
format_fa_list
}
\examples{
FALSE
}
