% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_page.R
\name{make_page}
\alias{make_page}
\title{Title}
\usage{
make_page(.navbar = el_navbar(), .main = el_main(), .load_wordcloud2 = TRUE)
}
\arguments{
\item{.navbar}{NULL}

\item{.main}{NULL}

\item{.load_wordcloud2}{NULL}
}
\description{
Title
}
\examples{
NULL
}
