% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_formatters.R
\name{rlist_c}
\alias{rlist_c}
\title{Title}
\usage{
rlist_c(.x, .fn_leaf_names = NULL, .fn_leaf_wrap = NULL, .fn_node_wrap = NULL)
}
\arguments{
\item{.x}{}

\item{.fn_leaf_names}{}

\item{.fn_leaf_wrap}{}

\item{.fn_node_wrap}{}
}
\value{
string
}
\description{
Title
}
\examples{
FALSE
}
